refactor: Extracted Address Struct to new file
