+++ entity/user.go
@@ -17,7 +17,7 @@ type User struct {
 	BirthDate *time.Time        `json:"birthdate"`
 	CreatedAt *time.Time        `json:"createdAt" db:"created_at"`
 	UpdatedAt *time.Time        `json:"updatedAt" db:"updated_at"`
-	Adresses  Address         `json:"addresses"`
+	Adresses  Address           `json:"addresses"`
 }
 
 func NewUser(name, uType, document, avatar_url, email, phone, pass, city, state string, birthdate *time.Time) *User {
@@ -37,24 +37,4 @@ func NewUser(name, uType, document, avatar_url, email, phone, pass, city, state 
 		BirthDate: birthdate,
 		Adresses:  address,
 	}
-}
-
-func NewAddress(userId uniqueEntityId.ID, city, state string) Address {
-	return Address{
-		ID:      uniqueEntityId.NewID(),
-		UserId:  userId,
-		Address: "",
-		City:    city,
-		State:   state,
-	}
-}
-
-type Address struct {
-	ID        uniqueEntityId.ID `json:"id" db:"id"`
-	UserId    uniqueEntityId.ID `json:"userId" db:"userId"`
-	Address   string            `json:"address" db:"address"`
-	City      string            `json:"city" db:"city"`
-	State     string            `json:"state" db:"state"`
-	Latitude  float64           `json:"latitude" db:"latitude"`
-	Logintute float64           `json:"longitute" db:"longitute"`
 }
